Restructure Packages
Note: as part of this refactor, the access to Machine::shutdown method was opened from protected to public. The server application doesn't have an API and isn't intended to have its code invoked from the outside, so I don't think that exposing things like this should really be an issue.